+++ .prettierignore
@@ -1,0 +1,6 @@
+node_modules
+lib
+.env
+package-lock.json
+package.json
+tsconfig.json
