+++ .prettierrc.json
@@ -1,1 +1,1 @@
-{}
+{ "trailingComma": "es5", "tabWidth": 4, "semi": true, "singleQuote": true }
